Move from jest to vitest
It took me the entire day to figure out how to mock my
handleAvatarUpload function. It turned out that my code was not a
problem, the test runner was (or, specifically, using imports
instead of requires in it). Changing the test runner to vitest was all
it took. Plus, vitest is MUCH faster.